+++ src/modules/emote_autocomplete/components/Emotes.jsx
@@ -2,7 +2,6 @@ import React, {useEffect, useRef, useState} from 'react';
 import {AutoSizer, List} from 'react-virtualized';
 import useChatInput from '../hooks/ChatInput.jsx';
 import styles from './Emotes.module.css';
-import useRowNavigation from '../hooks/RowKeyboardNavigation.jsx';
 import keyCodes from '../../../utils/keycodes.js';
 import renderRow from './EmoteRow.jsx';
 import EmotesHeader from './EmotesHeader.jsx';
@@ -26,14 +25,29 @@ function calcMaxHeight(length) {
   return currentHeight;
 }
 
-let navigationCallback;
-function setNavigationCallback(callback) {
-  navigationCallback = callback;
+function travelUp(currentSelection, rowCount) {
+  const newSelection = currentSelection - 1;
+
+  if (newSelection <= -1) {
+    return rowCount - 1;
+  }
+
+  return newSelection;
+}
+
+function travelDown(currentSelection, rowCount) {
+  const newSelection = currentSelection + 1;
+
+  if (newSelection > rowCount - 1) {
+    return 0;
+  }
+
+  return newSelection;
 }
 
 export default function Emotes({chatInputElement, repositionPopover, onComplete, getAutocomplete}) {
   const [emotes, setEmotes] = useChatInput(chatInputElement, getAutocomplete);
-  const [selected, setSelected] = useRowNavigation(setNavigationCallback, emotes.length);
+  const [selected, setSelected] = useState(0);
   const [open, setOpen] = useState(true);
   const localRef = useRef(null);
 
@@ -68,16 +82,33 @@ export default function Emotes({chatInputElement, repositionPopover, onComplete,
         return;
       }
 
-      setOpen(true);
-
-      if (event.key === keyCodes.Enter || event.key === keyCodes.Tab) {
-        event.stopPropagation();
-        event.preventDefault();
-        handleAutocomplete(emotes[selected]);
-        return;
+      switch (event.key) {
+        case keyCodes.Enter:
+        case keyCodes.Tab:
+          event.preventDefault();
+          event.stopPropagation();
+          handleAutocomplete(emotes[selected]);
+          break;
+        case keyCodes.End:
+          event.preventDefault();
+          setSelected(emotes.length);
+          break;
+        case keyCodes.Home:
+          event.preventDefault();
+          setSelected(0);
+          break;
+        case keyCodes.ArrowUp:
+          event.preventDefault();
+          setSelected(travelUp(selected, emotes.length));
+          break;
+        case keyCodes.ArrowDown:
+          event.preventDefault();
+          setSelected(travelDown(selected, emotes.length));
+          break;
+        default:
+          setOpen(true);
+          setSelected(0);
       }
-
-      navigationCallback(event);
     }
 
     window.addEventListener('keydown', keydownCallback, true);
